+++ app/src/test/admin/components/ModellingGroups/SingleGroup/GroupAdminContentTests.tsx
@@ -1,11 +1,11 @@
 import * as React from "react";
 import { expect } from "chai";
 import { shallow } from "enzyme";
-import { GroupAdminContentComponent } from "../../../../../main/admin/components/ModellingGroups/SingleGroup/Admin/GroupAdminContent";
+import { GroupAdminContentComponent } from "../../../../../main/admin/components/ModellingGroups/SingleGroup/Members/GroupAdminContent";
 import { mockModellingGroupDetails, mockUser } from "../../../../mocks/mockModels";
 import { alt } from "../../../../../main/shared/alt";
 import { ListOfUsers } from "../../../../../main/admin/components/ModellingGroups/ListOfUsers";
-import { AddMember } from "../../../../../main/admin/components/ModellingGroups/SingleGroup/Admin/AddMember";
+import { AddMember } from "../../../../../main/admin/components/ModellingGroups/SingleGroup/Members/AddMember";
 
 describe("GroupAdminContent", () => {
